lots of refactoring, getting ready to make Maps editable
the main outstanding pathfinding bug was the stupid .Net SortedList,
which just stinks. So I added SortQueue[T] which seems to work and
seems to be performant for reasonable lists of values. It’s biased so
that later nodes with equivalent heuristics pop first, which seems to
encourage straight-line seeks